Compatibility remove index with Uuid / Ulid
This does not work correctly with a UUID or ULID object as the primary key.

To index a document, no problem, Meilisearch does convert the ObjectId into a character string.

On the other hand, to delete a resource from MeiliSearch, you must indicate the ObjectId in character string in the url of the DELETE request.

While waiting for the refactoring on the objectID which will surely correct this bug (# 58).

My fix while waiting for the refacto of the objectId is to indicate the ObjectID property of the normalized object instead of entity-> getID ().