Test case for contents matching object-ID
If the contents of a post happen to match an object-ID of any post, only
that post is returned N times. This extends to arbitrary entities when a
single indexed field contains a value that happens to match an
object-ID.
This is due to `findOneBy` expanding the `IN(?)`-criteria from a single
parameter to a list of parameters `IN(?,?,?)` when the value of the
`by`-condition is an array.

However, we expect the search to still return all different objects by
ID, instead of repeating the same object with the ID that happens to
match any other indexed property.