Pass a scalar object-ID to findOneBy
The search service falsely assumes that what it gets back in the
`hits`-array are object-IDs, when it actually contains arrays
holding the indexed data of the results.

When passing this array to
   findOneBy(['id' => $theArray])
Doctrine expands the `WHERE id IN(?)`-condition to a list of parameters:
`WHERE id IN(?,?,?)`, binding the parameters to the values of the
indexed properties.

If such a property happens to match an object-ID, e.g. if the indexed
data is:
```
id: 2
content: 1
```
this clause becomes `WHERE id IN(2, 1)`, and returns the first object it
finds multiple times.

By ensuring that we always pass the object-ID rather than the full hit,
this expansion is avoided and the correct entities are collected.